pyproject.toml
[project]
name = "timetable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.0.0",
    "ical>=12.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
